refactor(api): move the JWT middleware out of the composition root
`server/index.ts` cannot be type-aware linted. It imports every route module,
and giving a file with that import fan-in full type information exhausts the
heap: eslint died with SIGABRT after 91s at 4GB and after 189s at 6GB, needed
8GB to finish at all, and took ~3min there. So every commit that staged this
file failed the lint step, and CI's `eslint .` only survived it by amortizing
the program across a thousand other files.

Splitting the file does not fix that, and the measurement says so: flattening
the 157-line chained `.route()` accumulation into 81 separate statements still
OOMed at 4GB after 147s. The cost is the fan-in, not the length -- any file that
mounts 81 routers imports all 81 of them. What does fix it is not asking for type
information here at all: with type-aware rules off, the entry lints in 2.7s at
4GB.

The trade is that the typed rules stop covering this file, so nothing
security-bearing may live in it -- which is why jwtAuthMiddleware moves to
server/lib/middleware/jwt-auth.ts, joining the other middlewares. It is the
cross-tenant guard, the password-change token invalidation and the ScopedDB
injection; it belongs under no-floating-promises. Linting it for the first time
immediately surfaced four warnings that were unreachable before (an `any` from
JSON.parse on the JWT header, and an always-truthy condition).

Registration order is unchanged (`app.use('*', jwtAuthMiddleware)` sits in the
same place between the host-derived middlewares and the tenant-scoped ones) and
tests/unit/platform/middleware-order.spec.ts still pins it. The two agent
allowlist specs now import the middleware module instead of server/index.ts, so
they no longer drag in the whole app graph to check a path predicate.

With the entry file exempt, the hook's lint step no longer needs its own 8GB
heap and runs inside the global 4GB ceiling -- four files including this one lint
in 9.3s.

Verified: type-check (app + api), the three specs that pin the middleware (20
tests), and lint:tenant-scope / naming / tests / status-literals / timestamps /
migrefs / filesize / deadcode.
