+++ scripts/file-size-baseline.json
@@ -7,13 +7,13 @@
   "server/durable-objects/inspection-doc.ts": 912,
   "app/routes/inspections.tsx": 891,
   "server/lib/collab/results-doc.ts": 874,
-  "server/index.ts": 871,
   "server/api/sms.ts": 824,
   "app/components/portal/sections/ReportView.tsx": 776,
   "app/routes/settings-communication.tsx": 776,
   "server/services/inspection.service.ts": 752,
   "app/routes/settings-communication-templates.tsx": 724,
   "app/routes/template-edit.tsx": 719,
+  "server/index.ts": 716,
   "app/components/media-studio/PhotoAnnotator.tsx": 692,
   "server/api/inspections/report-delivery.ts": 667,
   "server/services/inspection/inspection-analytics.service.ts": 667,
