+++ tests/unit/agent/agent-login.spec.ts
@@ -246,14 +246,14 @@ describe('Agent password + magic-link login (core /agent-login front door)', () 
 /**
  * Allowlist guard. Both entry points are UNAUTHENTICATED (no session, no
  * report token) and must bypass the global JWT middleware
- * (`isAgentPublic` in server/index.ts) so a stale/expired session cookie on
- * the visiting browser can never 401 either request. Exercises the REAL
- * exported jwtAuthMiddleware directly, mirroring
+ * (`isAgentPublic` in server/lib/middleware/jwt-auth.ts) so a stale/expired
+ * session cookie on the visiting browser can never 401 either request.
+ * Exercises the REAL exported jwtAuthMiddleware directly, mirroring
  * tests/unit/agent/magic-login.spec.ts's own allowlist guard.
  */
 describe('Agent login/login-link — jwtAuthMiddleware allowlist', { timeout: 30_000 }, () => {
     it('bypasses JWT verification for both entry points regardless of a stale Bearer token', async () => {
-        const { jwtAuthMiddleware } = await import('../../../server/index');
+        const { jwtAuthMiddleware } = await import('../../../server/lib/middleware/jwt-auth');
         for (const path of ['/api/agent/login', '/api/agent/login-link']) {
             const next = vi.fn(async () => {});
             const fakeContext = {
